refactor(pipeline): extract shared base for NativeToPlace and PlaceToMove stages
_NativeToPlaceBase (template-method) holds the shared squin→native→place
sequence. Subclasses override three hooks:

  _pre_native_rewrites(mt, out, no_raise) — logical: ValidationSuite + callgraph
  _post_unroll_validation(out)            — physical: terminal-measure check
  _lower_qubits(out)                      — physical/logical qubit-lowering rewrites

_PlaceToMove replaces the two identical PlaceToMove classes; the only
behavioural difference (InsertInitialize) is now a single bool flag.

_default_merge_heuristic is also deduplicated into base.py.
